# perf(nn/stateless): cache intermediate submodules