## Description

Describe your changes in detail.

## Motivation and Context

Why is this change required? What problem does it solve?
If it fixes an open issue, please link to the issue here.
You can use the syntax `close #15213` if this solves the issue #15213

- [ ] I have raised an issue to propose this change ([required](https://github.com/metaopt/torchopt/issues) for new features and bug fixes)

Cache intermediate submodules for better performance.

Before:

```python
In [1]: import torchopt

In [2]: from torchvision.models import resnet152

In [3]: model = resnet152()

In [4]: params = dict(model.named_parameters())

In [5]: %timeit torchopt.nn.swap_state(model, params)
1.16 ms ± 6.23 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

After:

```python
In [1]: import torchopt

In [2]: from torchvision.models import resnet152

In [3]: model = resnet152()

In [4]: params = dict(model.named_parameters())

In [5]: %timeit torchopt.nn.swap_state(model, params)
677 µs ± 4.32 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

## Checklist

Go over all the following points, and put an `x` in all the boxes that apply.
If you are unsure about any of these, don't hesitate to ask. We are here to help!

- [X] I have read the [CONTRIBUTION](https://torchopt.readthedocs.io/en/latest/developer/contributing.html) guide. (**required**)
- [X] My change requires a change to the documentation.
- [X] I have updated the tests accordingly. (*required for a bug fix or a new feature*)
- [X] I have updated the documentation accordingly.
- [X] I have reformatted the code using `make format`. (**required**)
- [X] I have checked the code using `make lint`. (**required**)
- [X] I have ensured `make test` pass. (**required**)